navbars to all
some have issues with bootstrap and CSS... however the same CSS is used everywhere.
some files are using bootstrap 4, and some are using bootstrap 5, that could be another issue.
also, issues like working for some time and then suddenly not working have occurred and some time change in the browser has resulted in a slight change in display. can look into it tomorrow........